flask_app/requirements.txt
boto3==1.19.12
botocore==1.22.12
click==8.0.3
cycler==0.11.0
flask-cors==3.0.10
flask==2.0.2
itsdangerous==2.0.1
jinja2==3.0.2
jmespath==0.10.0
kiwisolver==1.3.2
markupsafe==2.0.1
matplotlib==3.4.3
mpmath==1.2.1
numpy==1.21.4
pillow==8.4.0
pyparsing==3.0.4
python-dateutil==2.8.2
python-dotenv==0.19.1
s3transfer==0.5.0
six==1.16.0
sympy==1.9
urllib3==1.26.7
uwsgi==2.0.20
werkzeug==2.0.2
